@@ -1,33 +1,44 @@
 [tool.poetry]
 name = "datar"
-version = "0.5.5"
-description = "Port of dplyr and other related R packages in python, using pipda."
+version = "0.15.6"
+description = "A Grammar of Data Manipulation in python"
 authors = ["pwwang <pwwang@pwwang.com>"]
-readme = "README.md"
 license = "MIT"
+readme = "README.md"
 homepage = "https://github.com/pwwang/datar"
 repository = "https://github.com/pwwang/datar"
 
 [tool.poetry.dependencies]
-python = "^3.7.1" # align with pandas
-pipda = "^0.4.5"
-diot = "^0.1.1"
-# pipda already requires the following
-# executing = "*"
-# varname = "*"
-pandas = "^1.2"
-scipy = { version = "^1.6", optional = true }
-wcwidth = { version = "^0.2", optional = true }
-python-slugify = {version = "^5.0.2", optional = true}
+python = "^3.8"
+simplug = "^0.4"
+pipda = "^0.13.1"
+python-simpleconf = {version = "^0.6", extras = ["toml"]}
+datar-numpy = {version = "^0.3.4", optional = true}
+datar-pandas = {version = "^0.5.5", optional = true}
+# datar-polars = {version = "^0.0.0", optional = true}
+datar-arrow = {version = "^0.1", optional = true}
 
-[tool.poetry.dev-dependencies]
-pytest = "*"
-pytest-cov = "*"
-pytest-xdist = "^2.4.0"
+[tool.poetry.build]
+generate-setup-file = true
+
+[tool.poetry.extras]
+numpy = ["datar-numpy"]
+pandas = ["datar-pandas"]
+arrow = ["datar-arrow"]
+# modin = ["datar-pandas"]
+# polars = ["datar-polars"]
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.1"
+pytest-cov = "^4"
+six = "^1.16"
+numpy = "^1.17"
+python-slugify = "^8"
+
 
 [build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
 
 [tool.mypy]
 ignore_missing_imports = true
@@ -37,9 +48,9 @@ show_error_codes = true
 strict_optional = false
 
 [tool.pytest.ini_options]
-addopts = "-vv -n auto --cov-config=.coveragerc --cov=datar --cov-report xml:cov.xml --cov-report term-missing"
+addopts = "-vv -p no:asyncio --tb=short --cov-config=.coveragerc --cov=datar --cov-report xml:cov.xml --cov-report term-missing"
 filterwarnings = [
-    "error"
+    # "error"
 ]
 console_output_style = "progress"
 junit_family = "xunit1"
