@@ -1,10 +1,13 @@
 """Provide options"""
+from __future__ import annotations
 
-from typing import Any, Generator, Mapping, Union, Callable
+from typing import Any, Generator, Mapping
 from contextlib import contextmanager
 
 from diot import Diot
+from simpleconf import Config
 
+from .defaults import OPTION_FILE_CWD, OPTION_FILE_HOME
 
 _key_transform = lambda key: key.replace("_", ".")
 _dict_transform_back = lambda dic: {
@@ -12,27 +15,24 @@ _dict_transform_back = lambda dic: {
 }
 
 OPTIONS = Diot(
-    # Whether use 0-based numbers when index is involved, acts similar like R
-    # Otherwise, like python
-    index_base_0=False,
-    # Whether which, which_min, which_max is 0-based
-    which_base_0=True,
-    dplyr_summarise_inform=True,
-    # whether warn about importing functions that override builtin ones.
-    warn_builtin_names=True,
-    add_option=True,
-    # allow 'a.b' to access 'a_b'
+    Config.load(
+        {
+            # Do we allow to use conflict names directly?
+            "allow_conflict_names": False,
+            # Disable some installed backends
+            "backends": [],
+        },
+        OPTION_FILE_HOME,
+        OPTION_FILE_CWD,
+        ignore_nonexist=True,
+    ),
     diot_transform=_key_transform,
-)
-
-OPTION_CALLBACKS = Diot(
-    # allow 'a.b' to access 'a_b'
-    diot_transform=_key_transform
 )
 
 
 def options(
-    *args: Union[str, Mapping[str, Any]],
+    *args: str | Mapping[str, Any],
+    _return: bool = None,
     **kwargs: Any,
 ) -> Mapping[str, Any]:
     """Allow the user to set and examine a variety of global options
@@ -40,8 +40,14 @@ def options(
     Args:
         *args: Names of options to return
         **kwargs: name-value pair to create/set an option
+        _return: Whether return the options.
+            If `None`, turned to `True` when option names provided in `args`.
+
+    Returns:
+        The options before updating if `_return` is `True`.
     """
-    if not args and not kwargs:
+    if not args and not kwargs and (_return is None or _return is True):
+        # Make sure the options won't be changed
         return OPTIONS.copy()
 
     names = [arg.replace(".", "_") for arg in args if isinstance(arg, str)]
@@ -51,23 +57,25 @@ def options(
             pairs.update(_dict_transform_back(arg))
     pairs.update(_dict_transform_back(kwargs))
 
-    out = Diot(
-        {
-            name: value
-            for name, value in OPTIONS.items()
-            if name in names or name in pairs
-        },
-        diot_transform=_key_transform,
-    )
+    out = None
+    if _return is None:
+        _return = names
+
+    if _return:
+        out = Diot(
+            {
+                name: value
+                for name, value in OPTIONS.items()
+                if name in names or name in pairs
+            },
+            diot_transform=_key_transform,
+        )
 
     for key, val in pairs.items():
         oldval = OPTIONS[key]
         if oldval == val:
             continue
         OPTIONS[key] = val
-        callback = OPTION_CALLBACKS.get(key)
-        if callable(callback):
-            callback(val)
 
     return out
 
@@ -85,8 +93,8 @@ def options_context(**kwargs: Any) -> Generator:
 
 
 def get_option(x: str, default: Any = None) -> Any:
-    """Get the current value set for option ‘x’,
-    or ‘default’ (which defaults to ‘NULL’) if the option is unset.
+    """Get the current value set for option `x`,
+    or `default` (which defaults to `NULL`) if the option is unset.
 
     Args:
         x: The name of the option
@@ -95,9 +103,11 @@ def get_option(x: str, default: Any = None) -> Any:
     return OPTIONS.get(x, default)
 
 
-def add_option(x: str, default: Any = None, callback: Callable = None) -> None:
-    """Add an option"""
-    OPTIONS[x] = default
-    if callback:
-        OPTION_CALLBACKS[x] = callback
-        callback(default)
+def add_option(x: str, default: Any = None) -> None:
+    """Add an option
+
+    Args:
+        x: The name of the option
+        default: The default value if `x` is unset
+    """
+    OPTIONS.setdefault(x, default)
